requirements.txt
bcrypt==4.1.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
Flask==3.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.0.1
zipp==3.17.0
